temporal/worker/requirements.txt
temporalio==1.2.0
asyncio==3.4.3
requests==2.31.0
pydantic==1.10.9
python-dotenv==1.0.0
cryptography==41.0.1